fix(editor-state-testing): avoid implicit directive parsing
