+++ packages/editor/packages/editor-state-testing/src/index.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import { createMockCodeBlock, createMockEventDispatcher, createMockState, createMockViewport } from './index';
 
 describe('editor-state testing utilities', () => {
-	it('creates code blocks with derived grid, cursor, module id, and parsed directives', () => {
+	it('creates code blocks with derived grid, cursor, module id, and empty parsed directives', () => {
 		const block = createMockCodeBlock({
 			x: 16,
 			y: 32,
@@ -16,21 +16,31 @@ describe('editor-state testing utilities', () => {
 		expect(block.gridY).toBe(2);
 		expect(block.cursor).toEqual({ col: 0, row: 0, x: 56, y: 20 });
 		expect(block.moduleId).toBe('synth');
+		expect(block.parsedDirectives).toEqual([]);
+	});
+
+	it('allows tests to provide directive records explicitly', () => {
+		const block = createMockCodeBlock({
+			code: ['module synth', '; @home', 'moduleEnd'],
+			parsedDirectives: [
+				{
+					prefix: '@',
+					name: 'home',
+					args: [],
+					rawRow: 1,
+					sourceLine: '; @home',
+					isTrailing: false,
+				},
+			],
+		});
+
 		expect(block.parsedDirectives).toEqual([
 			{
 				prefix: '@',
 				name: 'home',
 				args: [],
 				rawRow: 1,
-				sourceLine: '; @home @favorite',
-				isTrailing: false,
-			},
-			{
-				prefix: '@',
-				name: 'favorite',
-				args: [],
-				rawRow: 1,
-				sourceLine: '; @home @favorite',
+				sourceLine: '; @home',
 				isTrailing: false,
 			},
 		]);
